infrastructure/pyproject.toml
[project]
name = "infrastructure"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "faiss-cpu>=1.12.0",
  "fastapi>=0.119.0",
  "flashinfer-python>=0.4.1",
  "httpx>=0.27.2",
  "numpy>=2.2.6",
  "orjson>=3.11.3",
  "pillow>=12.0.0",
  "xformers>=0.0.33.dev1091",
]

[tool.uv]
prerelease = "allow"
